backend/requirements.txt
blinker==1.9.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
Werkzeug==3.1.3
openai
python-dotenv==1.2.1
pytest==8.4.2
numpy
gunicorn
